Problem: unnecessary items in chain-core (fixes #1367)
Solution: removed the no longer existant parameters as per the revised staking:
https://github.com/crypto-com/chain-docs/blob/master/docs/modules/punishment.md
https://github.com/crypto-com/chain-docs/blob/master/docs/getting-started/reward-and-punishments.md#punishments-network-parameters-for-slashing
