+++ chain-abci/tests/abci_app.rs
@@ -159,14 +159,12 @@ fn get_dummy_network_params() -> NetworkParameters {
         required_council_node_stake: Coin::unit(),
         unbonding_period: 86400,
         jailing_config: JailingParameters {
-            jail_duration: 86400,
             block_signing_window: 100,
             missed_block_threshold: 50,
         },
         slashing_config: SlashingParameters {
             liveness_slash_percent: SlashRatio::from_str("0.1").unwrap(),
             byzantine_slash_percent: SlashRatio::from_str("0.2").unwrap(),
-            slash_wait_period: 10800,
         },
         rewards_config: RewardsParameters {
             monetary_expansion_cap: Coin::zero(),
@@ -354,14 +352,12 @@ fn init_chain_panics_with_different_app_hash() {
         required_council_node_stake: Coin::unit(),
         unbonding_period: 1,
         jailing_config: JailingParameters {
-            jail_duration: 86400,
             block_signing_window: 100,
             missed_block_threshold: 50,
         },
         slashing_config: SlashingParameters {
             liveness_slash_percent: SlashRatio::from_str("0.1").unwrap(),
             byzantine_slash_percent: SlashRatio::from_str("0.2").unwrap(),
-            slash_wait_period: 10800,
         },
         rewards_config: RewardsParameters {
             monetary_expansion_cap: expansion_cap,
