+++ chain-core/src/init/config.rs
@@ -25,7 +25,6 @@ pub enum DistributionError {
     InvalidValidatorAccount,
     NoValidators,
     InvalidVotingPower,
-    InvalidPunishmentConfiguration,
     InvalidRewardsParamter(&'static str),
 }
 
@@ -60,9 +59,6 @@ impl fmt::Display for DistributionError {
             DistributionError::InvalidVotingPower => {
                 write!(f, "Invalid voting power")
             },
-            DistributionError::InvalidPunishmentConfiguration => {
-                write!(f, "Invalid punishment configuration (maybe slash_wait_period >= jail_duration or jail_duration > unbond period)")
-            }
             DistributionError::InvalidRewardsParamter(err) => {
                 write!(f, "Invalid rewards parameters: {}", err)
             }
@@ -220,16 +216,6 @@ impl InitConfig {
             Err(e) => {
                 return Err(DistributionError::DistributionCoinError(e));
             }
-        }
-
-        if self.network_params.slashing_config.slash_wait_period
-            >= self.network_params.jailing_config.jail_duration
-        {
-            return Err(DistributionError::InvalidPunishmentConfiguration);
-        }
-
-        if self.network_params.jailing_config.jail_duration > self.network_params.unbonding_period {
-            return Err(DistributionError::InvalidPunishmentConfiguration);
         }
 
         let accounts = self.get_account(genesis_time);
