+++ chain-core/tests/verify_config.rs
@@ -64,14 +64,12 @@ fn test_verify_test_example_snapshot() {
         required_council_node_stake: Coin::new(5000_0000_0000_0000).unwrap(),
         unbonding_period: 86400,
         jailing_config: JailingParameters {
-            jail_duration: 86400,
             block_signing_window: 100,
             missed_block_threshold: 50,
         },
         slashing_config: SlashingParameters {
             liveness_slash_percent: SlashRatio::from_str("0.1").unwrap(),
             byzantine_slash_percent: SlashRatio::from_str("0.2").unwrap(),
-            slash_wait_period: 10800,
         },
         rewards_config: RewardsParameters {
             monetary_expansion_cap: expansion_cap,
