+++ client-common/src/tendermint/mock.rs
@@ -57,14 +57,12 @@ const DEFAULT_GENESIS_JSON: &str = r#"{
             "required_council_node_stake": "1250000000000000000",
             "unbonding_period": 86400,
             "jailing_config": {
-                "jail_duration": 86400,
                 "block_signing_window": 100,
                 "missed_block_threshold": 50
             },
             "slashing_config": {
                 "liveness_slash_percent": "0.100",
-                "byzantine_slash_percent": "0.200",
-                "slash_wait_period": 10800
+                "byzantine_slash_percent": "0.200"
             },
             "rewards_config": {
                 "monetary_expansion_cap": "6250000000000000000",
