+++ dev-utils/example-dev-conf.json
@@ -6,14 +6,12 @@
     "unbonding_period": 86400,
     "required_council_node_stake": "1250000000000000000",
     "jailing_config": {
-        "jail_duration": 86400,
         "block_signing_window": 100,
         "missed_block_threshold": 50
     },
     "slashing_config": {
         "liveness_slash_percent": "0.1",
-        "byzantine_slash_percent": "0.2",
-        "slash_wait_period": 10800
+        "byzantine_slash_percent": "0.2"
     },
     "rewards_config": {
         "monetary_expansion_cap": "6250000000000000000",
