+++ dev-utils/src/commands/genesis_dev_config.rs
@@ -37,14 +37,12 @@ impl GenesisDevConfig {
             unbonding_period: 86400,
             required_council_node_stake: Coin::new(1_250_000_000_000_000_000).unwrap(),
             jailing_config: JailingParameters {
-                jail_duration: 86400,
                 block_signing_window: 100,
                 missed_block_threshold: 50,
             },
             slashing_config: SlashingParameters {
                 liveness_slash_percent: SlashRatio::from_str("0.1").unwrap(),
                 byzantine_slash_percent: SlashRatio::from_str("0.2").unwrap(),
-                slash_wait_period: 10800,
             },
             rewards_config: RewardsParameters {
                 monetary_expansion_cap: expansion_cap,
